🧹 Move more logic into the invitation plugin
The less logic that lives in the component hierarchy the better.

We want to delegate as much domain complexity as possible to the plugin handler.
This allows us to focus on view concerns and dispatching actions, but not worry about
the domain logic from the view code.

We can focus on what we want to achieve and ignore the how. This unbraids a few concerns.